+++ src/app/app.component.ts
@@ -11,7 +11,7 @@ import { CalendarPage } from './pages/calendar/calendar';
 export class AppComponent {
 	@ViewChild(Nav) nav: Nav;
 
-	rootPage = HomePage;
+	rootPage = AmapPage;
 
 	pages: any[];